Moved Kernel components to separate namespace
