+++ src/Ninject.Web.AspNetCore/Components/OrderedDisposalStrategy.cs
@@ -2,8 +2,12 @@
 using Ninject.Activation.Strategies;
 using Ninject.Components;
 
-namespace Ninject.Web.AspNetCore
+namespace Ninject.Web.AspNetCore.Components
 {
+	/// <summary>
+	/// Replaces the default <see cref="Ninject.Activation.Strategies.DisposableStrategy"/> to integrate the <see cref="IDisposalManager"/>
+	/// mechanism for deterministic disposal order.
+	/// </summary>
 	public class OrderedDisposalStrategy : NinjectComponent, IActivationStrategy
 	{
 		public IDisposalManager DisposalManager { get; set; }
