+++ src/Ninject.Web.AspNetCore/NinjectServiceScope.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using Microsoft.Extensions.DependencyInjection;
 using Ninject.Infrastructure.Disposal;
+using Ninject.Web.AspNetCore.Components;
 
 namespace Ninject.Web.AspNetCore
 {
